+++ src-docs/src/views/form_layouts/form_rows.js
@@ -87,7 +87,7 @@ export default class extends Component {
 
   render() {
     return (
-      <EuiForm>
+      <EuiForm renderForm>
         <EuiFormRow label="Text field" helpText="I am some friendly help text.">
           <EuiFieldText name="first" />
         </EuiFormRow>